requirements.txt
streamlit
python-docx
google-genai
python-dotenv
tenacity
